Remove DNAConverterModel
